Updated fileActionHandler and abstract functions to support multi-file reading. Beware of breaking changes and patch your code!

+++ core/fileIO.php
@@ -15,19 +15,17 @@
 		decodeJsonFile - This returns an array based on JSON decode.
 		
 		createJsonFile - This creates a file based on the multidimensional array it receives, along with basic data such as name of file, 
-		and location to store it. CreateFile also has a built in auto-duplication process by creating the same file on other MetisDB nodes.
+		and location to store it.
 
-		readJsonFile - This returns the JSON Encoded file based on the file name and location. It creates a copy of the file with a unique
-		hash based on the date, time, I.P. address, and a random number above 100000. The chances of 100000 
-		individuals requesting the same file at the same second at the same computer is astronomical.
+		readJsonFile - This returns the JSON Encoded file based on the file name and location. 
 
 		updateJsonFile - This updates a JSON file (with input being an array of connections, the file name (hashed), and a multi-dimensional array.
 
 		deleteJsonFile - This deletes the stored JSON file. The requested file to be deleted can also be deleted from multiple MetisDB nodes.
 
-		replicator - This function replicates JSON files across a specific FTP and local nodes. You must specify the origin node (FTP or local), the 
-		nodes you wish to replicate to (this list of nodes must be individual items in an array), and the JSON file you wish to copy. It will copy the 
-		file with its existing name and overwrite the file if it already exists on that node.
+		replicator - This function replicates JSON files across a FTP and local nodes. You must specify the origin node (FTP or local), the 
+		nodes you wish to replicate to (this list of nodes must be individual items in an array), and the JSON files you wish to copy. It will copy the 
+		file(s) with its existing name and overwrite the file if it already exists on that node.
 
 		-----
 		
@@ -62,12 +60,18 @@
 		}
 	}
 	
-	function fileActionHandler(array $nodeArray, $fileName_NotHashed, $fileAction, array $fileContent_JsonArray = null){
-		$fileName = fileHashing($fileName_NotHashed); // Generate the hashed file name.
+	function fileActionHandler($nodeNum, array $files, $fileAction, array $fileContent_JsonArray = null){
 		global $nodeList; //Get the node list as a multi-dimensional array.
 		$fileContent = null; // fileContent variable used to return file content or errors.
 		$originalDirectory = getcwd(); // Get the current directory prior to moving to other directories.
-					
+		
+		if (count($files) == 1){ // If the amount of files to interact of with is one (a single file)
+			$multiFile = false; // We are not dealing with multiple files
+		}
+		else{ // If we are interacting with a multitude of files
+			$multiFile = true; // We are dealing with multiple files, so multiFile needs to be set to true
+		}
+		
 		if ($fileContent_JsonArray !== null){
 			$jsonData = json_encode($fileContent_JsonArray); // Encode the multidimensional array as JSON
 			if ($jsonData === false){
@@ -76,80 +80,78 @@
 		}
 		
 		if ($nodeList !== 1.01){ // If the nodeList was successfully fetched, we'll continue our file IO process
-		
-			foreach ($nodeArray as $key => $nodeNum){
+			if (atlasui_string_check($fileAction, array("r", "w", "a", "d")) !== false){
 				$nodeAddress = getNodeInfo($nodeList, $nodeNum, "Address");
 				$nodeType = getNodeInfo($nodeList, $nodeNum, "Node Type"); // Get the type of node that is being used.
 				$nodePreferentialLocation = getNodeInfo($nodeList, $nodeNum, "Preferential Location"); // Get the preferential location for the Node.
-			
-				if ($nodeType !== 1.03){ // If getting the node info did not fail
 				
-					if (atlasui_string_check($nodeType, array("ftp", "local")) == true){ // Make sure its either an FTP, MySQLi or local connection.
-							
-						if ($nodePreferentialLocation !== ""){ // If the Preferential Location is set.
+				if ($nodeAddress !== 1.03){ // If getting the node info did not fail
 				
-							if (strpos($nodeType, "local") !== false){
+					if (atlasui_string_check($nodeType, array("ftp", "local")) == true){ // Make sure its either an FTP or local connection.
+					
+						if ($nodePreferentialLocation !== ""){ // If the node's preferential location is not empty
+						
+							if ($nodeType == "local"){
 								navigateToLocalMetisData($nodeAddress, $nodePreferentialLocation); // Go to the data directory
 							}
-					
-							if (strpos(getcwd(), $nodePreferentialLocation) !== false){
-								if (atlasui_string_check($fileAction, array("r", "w", "a", "d")) !== false){
-									if ($nodeType == "ftp"){
-										$nodeConnection = establishConnection($nodeNum); // Establish a connection to create the file, since it's FTP and not local (which makes establishing a connection a logical step).
-
-										if (gettype($nodeConnection) !== "string"){							
-											if ($fileAction !== "d"){
-												$tempJsonFile = tmpfile(); // Create a temporary file to store the JSON info in.
+							elseif ($nodeType == "ftp"){
+								$nodeConnection = establishConnection($nodeNum); // Establish a connection to create the file, since it's FTP and not local (which makes establishing a connection a logical step).
+							}
 							
-												if ($fileAction !== "a"){
-													ftp_get($nodeConnection, $tempJsonFile, $fileName  . ".json", FTP_BINARY); // Read the file, saving it into the temporary JSON file.
-													fseek($tempJsonFile, 0); // Go to the first character in the file.
-													$fileContent =  file_get_contents($tempJsonFile); // Read the contents into fileContent.
-												}
-													
-												if ($fileAction == "r"){ // If the file action is to just read the contents
-													fclose($tempJsonFile); // Close the temporary file
-													return $fileContent; // Go ahead and return the file content, skipping over unnecessary writing and uploading that file action "w" and "a" would do.
-												}
-												else{
-													if($fileAction == "w"){ // If the file action is to update but NOT include the current file contents (essentially, overwrite the file).
-														$fileContent = fwrite($tempJsonFile, $jsonData); // Write the JSON data to the temporary file.
-													}
-													else{ // If the file action is to update AND append the contents
-														$fileContent = $fileContent . $jsonData; // Append contents 
-														fseek($tempJsonFile, 0); // Seek back to beginning of temporary file
-														$fileContent = fwrite($tempJsonFile, $jsonData); // Write the new data to the temporary file
-													}
+							if (($nodeType == "local") || (($nodeType == "ftp") && (gettype($nodeConnection) !== "string"))){		
+								foreach ($files as $fileName_NotHashed){
+									$fileName_Hashed = fileHashing($fileName_NotHashed); // Generate the hashed file name.		
+									$tempJsonFile = tmpfile(); // Create a temporary file to store the JSON info in.		
 							
-													fseek($tempJsonFile, 0); // Set the seek point to 0 so the entire file can be written to the FTP-based file.
-													ftp_put($nodeConnection, $fileName . ".json", $tempJsonFile, FTP_BINARY, 0); // Upload the new file contents as that JSON file
-													fclose($tempJsonFile); // Close the temporary file
-												}
-											}
-											else{ // If the file action is to delete
-												ftp_delete($establishConnection, $fileName); // Delete the file.
-											}
-										}
-										else{
-											return $nodeConnection; // As we don't know for certain if this is a connection or ftp_chdir issue, we'll return the error code carried over from establishConnection().
-										}
-									}
-									elseif ($nodeType == "local"){
-										if ($fileAction == "r"){
-											$fileContent = file_get_contents($fileName . ".json");
-											if ($fileContent == false){
+									if ($fileAction == ("r" || "a")){
+										if ($nodeType == "local"){
+											$thisFileContent = file_get_contents($fileName_Hashed . ".json"); // Read the file
+									
+											if ($thisFileContent == false){ // If we failed to fetch the file
 												return 2.05; // Return the error code #2.05.
 											}
 										}
-										elseif($fileAction !== "d"){
-											$fileHandler = fopen($fileName . ".json", "w+"); // Create a file handler (open the file with the requested fileAction and NO flags.
-							
-											if ($fileHandler !== false){
-												if ($fileAction == "a"){
-													$currentFileContent = readfile($fileHandler);
-													$jsonData = $currentFileContent . $jsonData;
+										elseif ($nodeType == "ftp"){
+											ftp_get($nodeConnection, $tempJsonFile, $fileName_Hashed  . ".json", FTP_BINARY); // Read the file, saving it into the temporary JSON file.
+											fseek($tempJsonFile, 0); // Go to the first character in the file.
+											$thisFileContent =  file_get_contents($tempJsonFile); // Read the contents into fileContent.
+										
+											fclose($tempJsonFile); // Close the temporary file
+										}
+									
+										if ($fileAction == "r"){								
+											if ($multiFile == false){ // If we are not fetching multiple files
+												$fileContent = $thisFileContent; // The file content we'll be returning is the content of this file
+											}
+											else{
+												$fileContent = $fileContent . "\"$fileName_NotHashed\" : " . $thisFileContent . ","; // Append the contents of this file into the fileContent, ensuring it is something like ["config"] : {FILE_CONTENT}
+											}
+										}
+										elseif ($fileAction == "a"){
+											$jsonData = $fileContent . $jsonData; // Append the JSON data to the current file content (appending)
+											if ($nodeType == "local"){
+												$fileHandler = fopen($fileName_Hashed . ".json", "w+"); // Create a file handler (open the file with the requested fileAction and NO flags.
+											
+												if ($fileHandler !== false){ // If we successfully opened the file for writing
+													fwrite($fileHandler, $jsonData); // Write the JSON data to the file.
+													fclose($fileHandler); // Close the file location.
 												}
-							
+												else{
+													return 2.05; // Return the error code #2.05.
+												}
+											}
+											else{
+												fseek($tempJsonFile, 0); // Seek back to beginning of temporary file
+												fwrite($tempJsonFile, $jsonData); // Write the new data to the temporary file
+												ftp_put($nodeConnection, $fileName_Hashed . ".json", $tempJsonFile, FTP_BINARY, 0); // Upload the new file contents as that JSON file
+											}
+										}
+									}
+									elseif ($fileAction == "w"){
+										if ($nodeType == "local"){ // If we are writing to a local file
+											$fileHandler = fopen($fileName_Hashed . ".json", "w+"); // Create a file handler (open the file with the requested fileAction and NO flags.
+										
+											if ($fileHandler !== false){ // If we successfully opened the file for writing
 												fwrite($fileHandler, $jsonData); // Write the JSON data to the file.
 												fclose($fileHandler); // Close the file location.
 											}
@@ -157,17 +159,31 @@
 												return 2.05; // Return the error code #2.05.
 											}
 										}
-										else{
+										else{ // If the file isn't local, it is accessible via FTP
+											fwrite($tempJsonFile, $jsonData); // Write the JSON data to the temporary file.
+											ftp_put($nodeConnection, $fileName_Hashed . ".json", $tempJsonFile, FTP_BINARY, 0); // Upload the new file contents as that JSON file
+										}								
+									}
+									else{ // If the fileAction == "d" (last option), then we're deleting files
+										if ($nodeType == "local"){
 											unlink($fileName);
 										}
+										else{
+											ftp_delete($establishConnection, $fileName); // Delete the file.
+										}
 									}
+								
+									fclose($tempJsonFile); // Close the temporary file before the end of the foreach
 								}
-						
-								chdir($originalDirectory);
-								return $fileContent;
 							}
 							else{
-								return false;
+								return $nodeConnection; // As we don't know for certain if this is a connection or ftp_chdir issue, we'll return the error code carried over from establishConnection().
+							}
+					
+							chdir($originalDirectory);
+							
+							if ($fileAction == "r"){
+								return "{" . substr($fileContent, 0, -1) . "}";
 							}
 						}
 						else{ // If the Preferential Location is blank...
@@ -198,15 +214,15 @@
 		return $jsonDecodedValue;
 	}
 	
-	function createJsonFile(array $nodeArray, $fileName_NotHashed, array $fileContent_JsonArray){
-		fileActionHandler($nodeArray, $fileName_NotHashed, "w", $fileContent_JsonArray);
+	function createJsonFile($nodeNum, array $files, array $fileContent_JsonArray){
+		fileActionHandler($nodeNum, $fileName_NotHashed, "w", $fileContent_JsonArray);
 	}
 	
-	function readJsonFile(array $nodeArray, $fileName_NotHashed){
-		return fileActionHandler($nodeArray, $fileName_NotHashed, "r");
+	function readJsonFile($nodeNum, array $files){ // This function is an abstraction for reading files and offers multi-file reading functionality
+		return fileActionHandler($nodeNum, $files, "r"); // Return the JSON file or error code from fileActionHandler
 	}
 	
-	function updateJsonFile(array $nodeArray, $fileName_NotHashed, array $fileContent_JsonArray, $fileAppend = false){
+	function updateJsonFile($nodeNum, $fileName_NotHashed, array $fileContent_JsonArray, $fileAppend = false){
 		if ($fileAppend == true){
 			$fileActionMode = "a";
 		}
@@ -214,28 +230,31 @@
 			$fileActionMode = "w";
 		}
 		
-		$updateResponse = fileActionHandler($nodeArray, $fileName_NotHashed, $fileActionMode, $fileContent_JsonArray);
+		$updateResponse = fileActionHandler($nodeNum, $fileName_NotHashed, $fileActionMode, $fileContent_JsonArray);
 		return $updateResponse;
 	}
 	
-	function deleteJsonFile(array $nodeArray, $fileName_NotHashed){
-		fileActionHandler($nodeArray, $fileName_NotHashed, "d");
+	function deleteJsonFile($nodeNum, $fileName_NotHashed){
+		fileActionHandler($nodeNum, $fileName_NotHashed, "d");
 	}
 	
-	function replicator(int $nodeSource, array $nodeDestinations_Array, $fileToReplicate_NameNotHashed){
+	function replicator($nodeSource, array $nodeDestinations_Array, array $files){
 		$nodeList = fetchNodeList(); // Fetch the Node List as a multidimensional array.
 		
-		if (isset($nodeList[$nodeSource])){ // Check if the Source / Origin Node exists.
-			$nodeOriginConnection = establishConnection($nodeSource); // Establish a connection with the origin node.
-			$nodeAddress = getNodeInfo($nodeList, $nodeSource, "Address"); // Get the source / origin node's full directory / address.
-			$nodeType = getNodeInfo($nodeList, $nodeSource, "Node Type"); // Get the source / origin node's type.
-			$nodePreferentialLocation = getNodeInfo($nodeList, $nodeSource, "Preferential Location"); // Get the source / origin node's preferential location.
+		if (isset($nodeList[$nodeSource])){ // Check if the Source / Origin Node exists.	
 			
-			$fileToReplicate_Content = readJsonFile(array($nodeList), $fileToReplicate_NameNotHashed);
-			
-			fileActionHandler($nodeDestinations_Array,  $fileToReplicate_NameNotHashed, "w", $fileToReplicate_Content);
+			foreach ($files as $fileName_NotHashed){ // Cycle through each file in the array
+				$fileContent_JsonFormat = readJsonFile($nodeSource, array($fileName_Hashed)); // Read the individual file we are replicating
+				$fileContent = decodeJsonFile($fileContent_JsonFormat); // Decode it into a multi-dimensional array for the purpose of using in createJsonFile
+				
+				if (gettype($fileContent) !== "int"){ // If the fileContent is not an int, generally meaning it isn't an error
+					foreach ($nodeDestinations_Array as $nodeDestination){ // Cycle through each node we are going to be replicating files to
+						createJsonFile($nodeDestination, array($fileName_Hashed), $fileContent); // Create (or overwrite) the file you are replicating on the node destination
+					}
+				}
+			}
 		}
-		else{
+		else{ // If the nodeSource does not exist
 			return "The requested origin node does not exist."; // If the origin node does NOT exist, output error.
 		}
 	}
